refactor: 디렉터리명을 database에서 data로 변경
